Checkpoint refactoring ParsedScv.
