+++ src/main/scala/org/broadinstitute/monster/clinvar/Constants.scala
@@ -1,11 +1,9 @@
 package org.broadinstitute.monster.clinvar
 
-import upack.{Msg, Str}
-
 /** Container for constants that we use across the ClinVar pipeline. */
 object Constants {
   /** Supported types of variants in VCVs and SCVs. */
-  val VariationTypes: Set[Msg] = Set("SimpleAllele", "Haplotype", "Genotype").map(Str)
+  val VariationTypes: Set[String] = Set("SimpleAllele", "Haplotype", "Genotype")
 
   /** Constant used to mark references to the MedGen database. */
   val MedGenKey: String = "MedGen"
